No longer need check_index(). Setters revamped. More setters, and existing ones can use getters to help with the job. Component setter.hpp no longer needed, so removed.
Certain getters, which we anticipate will only be of interest to C++ users rather than to Python users, actually return pointers instead of optionals. These are the occasional getters that try to return a particular TYPE of value from a variant. They may fail, if the value in the variant isn't of the requested type. In that case, a meaningful reference can't be returned. (Most of the getters return const or non-const references, as they should.) To be as consistent as possible with the other getters, we decided that returning a pointer (or nullptr if a value of the requested type isn't there) is best. That way, a caller can, as with the getters that return references, get access (now, via a pointer dereference) to the original object, not to a copy. And, to boot, we don't make a copy, which is generally a bad thing to do in this sort of place.