Began writing tests for Component.
Implemented a very flexible sort capability in Component, for derived classes.

Axes customization no longer needs its own sort; that happens automatically.

Any known (via keys()) vector in a Component-derived class is sorted (assuming finish() is called) automatically, by index and/or label if those are available. Variant and optional are handled in this respect as well. So, for example, a class might have a vector of variant, where one or more of the variant alternatives has index and/or label, either or both possible as optionals. The generic sort functionality does its level best to reasonably sort such an animal.

Made some of Component's private functionality public. This was partly motivated by wanting to test such things directly in Component's test codes, but that wasn't (and shouldn't be) the main motivation. Some things seem reasonable to make public, as codes outside of GNDStk might reasonably want to use those constructs in a Component object. I suspect that this is something we'll rethink and possibly change as GNDStk and our GNDS Standard Interface codes evolve.

Some cleanup in BodyText that allowed a few things to be removed, and some code to be shortened.

Tweaked some SFINAE constructs so code that uses it can be shortened.

Relaxed some SFINAE-enforced conditions in Component's getter() detail functions. Specifically, index and label no longer need to be exactly a size_t and a std::string, but just need to be convertible thereto. So, for example, index can be an int. Otherwise, these functions would have been unnecessarily awkward to use.