Cleanup/simplification of various .json files.
These are meant to make the .json files clearer, for the benefit of people who wish to make their own specs for GNDStk's code generator.

Renamed `__namespace__` to  `namespace`.
Renamed `__class__` to  `class`.
But don't actually need `class` at all.

Don't need `bodyText` if it's going to be `null`. (But it can still be given as `null`.)

Don't need `abstractNode`.
Don't need `description`.
Don't need `rootNode`.
Don't need `childUnique`.

Don't need `default` if it's going to be `null` anyway.
That is, `default` is needed only if there *is* a default.

Don't need `name` unless it's different from the current JSON key.