The json2class code now generates the key.hpp file. Made some largely superficial modifications to the key.hpp file as well.
Remark: it's entirely intentional that the `Meta` and `Child` objects in the `key.hpp` file are generic - not associated with a type (as with, for instance, `type{}/Meta<>("name")`). Metadata of the same name can end up having different types; for example, in one case `href` is an `XMLName`, while in another it's a `std::optional<XMLName>`. As for `Child` objects, we expect that they have a 1-to-1 correspondence with individual auto-generated classes, and thus we *could* put the class in as the type (as with `TheClass{}/Child<>("name")`). However, something else could be different - say, occuring once when it's a child of one particular parent class, or many times when it's a child of a different particular parent class. (Or perhaps it doesn't occur by itself at all, but as a variant alternative.) For this reason, and also to be consistent with how `Meta` objects are handled, all `Child` objects in `key.hpp` are made to be generic. In practice, this means that when they're used as-is in Core Interface queries, you'll get std::string (for metadata) and Node (for children).