Put in some constructs to simply SFINAE and if-constexprs in BodyText.
For as-yet-undetermined reasons, attempting to use BodyText::template supported<T> in the SFINAE for Component (which derives from BodyText) resulted in, apparently, some incorrect enables/disables that appeared to result in an infinite loop in one of the tests. So, we stuck with old tests based on isVoid and some other things.

Strangely,