Removed some unnecessary matieral from json2class.exe.
The removed code wrote output to a consolidated file. This was an early thing I did, to make it easy to examine *all* of the code the autogenerator produced, in one place. (As opposed to classes and such being placed into their own individual files.) The autogenerator is far enough along that I don't think it's necessary to keep this particular code around any longer.