More work on precision-related tests. (Not all in this commit.)
This work prompted a closer look at some of the convert() functions.

The convert() functions, after all, deal with (among other things) conversion between strings and values, where the values might be floating-point and thus the precision settings might be relevant.

Added a bit with regards to the convert() functions.

Was able to remove some Component detail:: material by formulating some of it in a more-flexible manner.