Circumvented pugixml's spurious newline printing. That's really the right thing to do (until/unless they fix it). XML, JSON, etc. are now more consistent - no spurious newline in XML. Updated test codes accordingly.
Fixed some fixmes.

General detail work, with the aim of tightening up the code.

Tweaked some comments.

More Node/Tree consolidation.
We use Node in most places, Tree only when necessary.

Removed some "inline"s when the function was a template anyway. Inline doesn't really mean "inline it" these days (compilers are good at making that decision), and it isn't necessary for header-only if the function is a template.