Added a whole bunch of "trying it out" GNDS 2.0 files.
These will be temporary, at least in this (try/ directory) location.

But I want to have these, now, in order to help ensure that upcoming code and spec changes have the effects -- and only the effects -- that I want them to have.

In the code generator, I replaced an assert with a meaningful error message, regarding times/occurrence. More work on the code generator will be forthcoming. (And hence one reason for putting in the slew of new files as described above.)