Added a minimal "mass" node to summary_tsl.json.
And, to use it, stipulated the "tsl" namespace for mass in scatteringAtom.

The above is meant as a fix, for now, of one of several ambiguities that the code generator reports during GNDS 2.0 code generation.

scatteringAtom in summary_tsl.json gives mass as a child node. However, prior to the above change, there was no "mass" in tsl (the present namespace), but a "mass" in each of summary_common.json and summary_pops.json.

There was no mechanism for the code generator to figure out which one was intended! So, I examined the tsl-*.xml files I have from GNDS 1.9, and observed that all <mass> nodes looked something like this:

     <mass value="90.436" unit="amu"/>

Judging from that, I placed a serviceable such "mass" class into "tsl".

The above fix is intended to be temporary, until the GNDS 2.0 spec clarifies precisely what "mass" in "scatteringAtom" should allow.