Disambiguated delayedNeutron's "product" child.
This is in: summary_fissionFragmentData.json.

The "product" node appears in two namespaces: pops and common. The pops version includes two attributes (label and pid), and no child nodes. The common version include the same two attributes, as well as children multiplicity, distribution, and outputChannel.

The "product" description in summary_fissionFragmentData.json says,

     "Product describing this delayed neutron.
      Includes average multiplicity and distribution info."

So, I infer that the common:: version, not the pops:: version, is required, and I specified "namespace" in order to make this happen.

Generated codes are updated accordingly.