Began an "incremental" GNDS 2.0 spec build, in standards/incremental.
Sorting of vectors in Component-derived classes is now optional, and OFF by default. That's really the right thing to do. For GNDS data (which is billed as being order-independent), it may seem convenient to sort vectors of objects according to an index or label metadatum in the object. However, (1) doing so takes time, and (2) we think that by default, at least, the order in which somebody enters such vector elements should be preserved.

Modified a couple of the GNDS 2.0 JSON specs, based on experience with trying to load GNDS 2.0 files.