Changed some terminology, updating test material appropriately.
Changing "comment" to "remark" may seem pointless at first glance, but serves a purpose. I'm currently preparing the way for some updating to how GNDStk handles certain special nodes - think CDATA and PCDATA in XML, as well as comments (in XML and elsewhere). Some modifications in the handling of special nodes will be important for the smooth handling of GNDS 2.0 files.

Throughout GNDStk, I'd been using the word "comment" in both a casual sense ("In this comment I'll describe such-and-such..."), and in a technical sense (the comment() function, comment nodes, etc.). To help make future grep'ing more useful, I decided to use "remark" for the casual sense, and keep "comment" for the technical sense. Henceforth, when I search for instances of the second case, I'll encounter less clutter.