Simple refactoring of some enums-namespace constructs.
Most enums were in the enums:: namespace. A couple of them relating to units were in a units/ directory. For now, I don't see any particular reason to categorize those enums as "units" enums. Other enums aren't categorized, after all. Splitting enums into subcategories might be useful if we had a huge number of enums, but we only have a handful. At the present time, then, I think that the simpler, slightly flatter factoring is just fine, and makes the files look just a bit more uniform.

I also changed around some enums-related ordering (of includes, and of tests given in CMake-related files) to be alphabetical, if for no other reason to be consistent and organized.