The code generator now (partially) handles enums.
At least, it accepts enumerator names, but doesn't fully deal with enumerator values yet.

Accepting enumerator names allowed us to fully autogenerate other codes that needed, for example, to #include or wrap (for Python) the enumerators.

Fully handling individual values (in a given enum class) is slated for the future. In this way, we'll be able to generate code that are currently handwritten in GNDStk/versions/GNDS-v2.0/GNDS/src/GNDS/v2.0/enums/.